taskmaster/api/api/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.4
packaging==24.1
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
Werkzeug==3.0.4